+++ CHANGELOG.md
@@ -23,6 +23,10 @@ To install or update CocoaPods see this [guide](http://docs.cocoapods.org/guides
   [#1933](https://github.com/CocoaPods/CocoaPods/issues/1933)
   [#1941](https://github.com/CocoaPods/CocoaPods/pull/1941)
 
+* Unset the `CDPATH` env variable before shelling-out on `prepare_command`  
+  [Marc Boquet](https://github.com/apalancat)
+  [#1943](https://github.com/CocoaPods/CocoaPods/pull/1943)
+
 ##### Bug Fixes
 
 * Fix :head option for Pods from HTTP source.